delete render.js,function rerenderPage moving to index.js
