+++ src/index.js
@@ -1,6 +1,22 @@
-import {rerenderPage} from "./render";
+import React from 'react';
+import ReactDOM from 'react-dom/client';
+import './index.css';
+import App from './App';
 import state from "./redux/state";
+import {addPost, updateNewPostText} from "./redux/state";
 
+
+const root = ReactDOM.createRoot(document.getElementById('root'));
+let rerenderPage = () => {
+
+    root.render(
+        <React.StrictMode>
+            <App state={state}
+                 addPost={addPost}
+                 updateNewPostText={updateNewPostText}/>
+        </React.StrictMode>
+    );
+}
 rerenderPage(state);
