--- src/render.js
@@ -1,22 +1,0 @@
-import React from 'react';
-import ReactDOM from 'react-dom/client';
-import './index.css';
-import App from './App';
-import {addPost, updateNewPostText} from "./redux/state";
-
-
-const root = ReactDOM.createRoot(document.getElementById('root'));
-export let rerenderPage = (state) => {
-
-    root.render(
-        <React.StrictMode>
-            <App state={state}
-                 addPost={addPost}
-                 updateNewPostText={updateNewPostText}/>
-        </React.StrictMode>
-    );
-}
-
-
-
-
